Add .editorconfig file
It's very important to remember the trailing line breaks in the md files
in the content dir. Pandoc is not putting artificial linebreaks in
between the files, but rather concat them together directly. Linebreaks
have significant meaning in markdown, and a missing linebreak before a
header will mean that it is no longer a header...

```
  some text from the previous file...
  # Header from the next file
```

Will be rendered as:

```html
  <p>some text from the previous file... # Header from the next file</p>
```

Instead of the intended:

```html
  <p>some text from the previous file...</p>
  <h1>Header from the next file</h1>
```

This is easily solved by configuring your editor to always add a
linebreak at the end of the file.